requirements.txt
Django==1.11
djangorestframework==3.6.2
gunicorn==19.7.1
presslabs-sugar==1.1.0
raven==6.0.0
